fix b_pow n ll

+++ cpp/lib/pow.hpp
@@ -1,10 +1,10 @@
 #include "base.hpp"
 
 template <typename T>
-T b_pow(const T& x, int n, const T& e = 1) {
+T b_pow(const T& x, ll n, const T& e = 1) {
   T m = e;
   T p = x;
-  int _n = n;
+  ll _n = n;
   while (_n > 0) {
     if (_n & 1) {
       m *= p;
@@ -15,7 +15,7 @@ T b_pow(const T& x, int n, const T& e = 1) {
   return m;
 }
 
-constexpr int const_pow_mod(int x, int n, int mod) {
+constexpr int const_pow_mod(int x, ll n, int mod) {
   return n == 0  ? 1
          : n & 1 ? const_pow_mod(ll(x) * x % mod, n >> 1, mod) * ll(x) % mod
                  : const_pow_mod(ll(x) * x % mod, n >> 1, mod);
